+++ db_api/models/user_health.py
@@ -10,7 +10,7 @@ class UserHealth(Base):
     __tablename__ = 'user_health'
 
     user_id = Column(BigInteger, primary_key=True)
-    health_id = Column(Integer, primary_key=True, autoincrement=True)
+    health_id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     type = Column(String, default='base')
     text = Column(String, default='')
