+++ db_api/models/user_limit.py
@@ -10,6 +10,6 @@ class UserLimit(Base):
     __tablename__ = 'user_limit'
 
     user_id = Column(BigInteger, primary_key=True)
-    limit_id = Column(Integer, primary_key=True, autoincrement=True)
+    limit_id = Column(BigInteger, primary_key=True, autoincrement=True)
     type = Column(String, default='base')
     text = Column(String, default='')
