+++ db_api/models/user_reg_page.py
@@ -10,7 +10,7 @@ class UserRegPage(Base):
     __tablename__ = 'user_reg_page'
 
     user_id = Column(BigInteger, primary_key=True)
-    reg_page_id = Column(Integer, primary_key=True, autoincrement=True)
+    reg_page_id = Column(BigInteger, primary_key=True, autoincrement=True)
 
     time = Column(DateTime, default=datetime.now, onupdate=datetime.now)
     reg_page_name = Column(String)
